Create SVM machine learning problem: Solved with data set from breast cancer  treatment
The code above corresponds to a machine learning problem particularly 'Support Vector Machine' with soft margin classifier involving a penalty parameter. The data set which has been loaded at the top of the program is a realistic set of breast cancer data and our task is to classify the data into two segments which has been formed in terms of a optimization problem. The solution has been obtained using the quadratic extension of Mehrotra's predictor corrector interior point algorithm.